# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flight Delay Risk API",
    "description": "Ranks scheduled flights by arrival-delay risk and returns a post-hoc calibrated estimate of 15+ minute delay, using only pre-flight information. Includes optional point-in-time weather and a European context layer.",
    "version": "1.5.0"
  },
  "paths": {
...
    "/predict": {
      "post": {
        "tags": [
          "prediction"
        ],
        "summary": "Predict",
        "operationId": "predict_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/predict/weather": {
      "post": {
        "tags": [
          "weather"
        ],
        "summary": "Predict With Weather",
        "description": "Return base prediction, point-in-time observations and paired weather delta.",
        "operationId": "predict_with_weather_predict_weather_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Predict With Weather Predict Weather Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/predict/european": {
      "post": {
        "tags": [
          "prediction"
        ],
        "summary": "Predict European",
        "operationId": "predict_european_predict_european_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EuropeanFlightInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EuropeanPredictionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/predict/batch": {
      "post": {
        "tags": [
          "prediction"
        ],
        "summary": "Predict Batch Endpoint",
        "operationId": "predict_batch_endpoint_predict_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchFlightInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPredictionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rank": {
      "post": {
        "tags": [
          "ranking"
        ],
        "summary": "Rank Flights Endpoint",
        "operationId": "rank_flights_endpoint_rank_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchFlightInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/flight": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Flight Report",
        "operationId": "flight_report_reports_flight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightReportInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/schedule": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Schedule Report",
        "operationId": "schedule_report_reports_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleReportInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchFlightInput": {
        "properties": {
          "flights": {
            "items": {
              "$ref": "#/components/schemas/FlightInput"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Flights"
          }
        },
        "type": "object",
        "required": [
          "flights"
        ],
        "title": "BatchFlightInput"
      },
      "BatchPredictionOutput": {
        "properties": {
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/PredictionOutput"
            },
            "type": "array",
            "title": "Predictions"
          }
        },
        "type": "object",
        "required": [
          "predictions"
        ],
        "title": "BatchPredictionOutput"
      },
...
      "EuropeanFlightInput": {
        "properties": {
          "airline": {
            "type": "string",
            "title": "Airline",
            "description": "European airline code, e.g. 'IB'."
          },
          "origin": {
            "type": "string",
            "title": "Origin",
            "description": "European origin airport IATA code, e.g. 'BCN'."
          },
          "destination": {
            "type": "string",
            "title": "Destination",
            "description": "European destination airport IATA code, e.g. 'AMS'."
          },
          "month": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Month"
          },
          "day_of_week": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Day Of Week"
          },
          "crs_dep_time": {
            "type": "integer",
            "maximum": 2400.0,
            "minimum": 0.0,
            "title": "Crs Dep Time"
          },
          "crs_arr_time": {
            "type": "integer",
            "maximum": 2400.0,
            "minimum": 0.0,
            "title": "Crs Arr Time"
          },
          "crs_elapsed_time": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Crs Elapsed Time"
          },
          "distance": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance",
            "description": "Optional manual distance in miles."
          }
        },
        "type": "object",
        "required": [
          "airline",
          "origin",
          "destination",
          "month",
          "day_of_week",
          "crs_dep_time",
          "crs_arr_time",
          "crs_elapsed_time"
        ],
        "title": "EuropeanFlightInput",
        "description": "Experimental European flight layer. Distance is auto-estimated if missing."
      },
      "EuropeanPredictionOutput": {
        "properties": {
          "delay_probability": {
            "type": "number",
            "title": "Delay Probability",
            "description": "Post-hoc calibrated estimate of a 15+ minute arrival delay."
          },
          "raw_model_score": {
            "type": "number",
            "title": "Raw Model Score",
            "description": "Uncalibrated score emitted by the selected classifier."
          },
          "calibration_method": {
            "type": "string",
            "title": "Calibration Method",
            "description": "Post-hoc calibration method applied to the raw score."
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level",
            "description": "'low', 'moderate', or 'high'."
          },
          "decision_threshold": {
            "type": "number",
            "title": "Decision Threshold",
            "description": "Probability threshold used for binary risk decisions."
          },
          "top_factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Factors",
            "description": "Non-causal schedule-context signals for this flight."
          },
          "local_contributions": {
            "items": {
              "$ref": "#/components/schemas/LocalContributionOutput"
            },
            "type": "array",
            "title": "Local Contributions",
            "description": "Signed local pre-calibration log-odds contributions from the selected model."
          },
          "explanation_scale": {
            "type": "string",
            "title": "Explanation Scale",
            "description": "Scale used by local model contributions.",
            "default": "log_odds_before_calibration"
          },
          "review_recommended": {
            "type": "boolean",
            "title": "Review Recommended",
            "description": "Whether the configured operational policy sends this flight to priority review.",
            "default": false
          },
          "operational_action": {
            "type": "string",
            "title": "Operational Action",
            "default": "standard_monitoring"
          },
          "policy_name": {
            "type": "string",
            "title": "Policy Name",
            "default": "fixed_threshold"
          },
          "policy_probability_cutoff": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Probability Cutoff"
          },
          "policy_capacity_fraction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Capacity Fraction"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "airline_name": {
            "type": "string",
            "title": "Airline Name"
          },
          "origin_label": {
            "type": "string",
            "title": "Origin Label"
          },
          "destination_label": {
            "type": "string",
            "title": "Destination Label"
          },
          "distance_miles": {
            "type": "number",
            "title": "Distance Miles"
          },
          "distance_source": {
            "type": "string",
            "title": "Distance Source"
          },
          "european_context": {
            "$ref": "#/components/schemas/EuropeanContextOutput"
          },
          "experimental": {
            "type": "boolean",
            "title": "Experimental",
            "default": true
          },
          "transfer_note": {
            "type": "string",
            "title": "Transfer Note"
          }
        },
        "type": "object",
        "required": [
          "delay_probability",
          "raw_model_score",
          "calibration_method",
          "risk_level",
          "decision_threshold",
          "top_factors",
          "region",
          "airline_name",
          "origin_label",
          "destination_label",
          "distance_miles",
          "distance_source",
          "european_context",
          "transfer_note"
        ],
        "title": "EuropeanPredictionOutput"
      },
      "FlightInput": {
        "properties": {
          "airline": {
            "type": "string",
            "title": "Airline",
            "description": "Operating/reporting carrier code, e.g. 'DL'.",
            "examples": [
              "DL"
            ]
          },
          "origin": {
            "type": "string",
            "title": "Origin",
            "description": "Origin airport IATA code, e.g. 'JFK'.",
            "examples": [
              "JFK"
            ]
          },
          "destination": {
            "type": "string",
            "title": "Destination",
            "description": "Destination airport IATA code, e.g. 'LAX'.",
            "examples": [
              "LAX"
            ]
          },
          "month": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Month",
            "description": "Scheduled month (1-12)."
          },
          "day_of_week": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Day Of Week",
            "description": "Scheduled day of week (1=Mon ... 7=Sun)."
          },
          "crs_dep_time": {
            "type": "integer",
            "maximum": 2400.0,
            "minimum": 0.0,
            "title": "Crs Dep Time",
            "description": "Scheduled departure time, HHMM format. 2400 is accepted as midnight."
          },
          "crs_arr_time": {
            "type": "integer",
            "maximum": 2400.0,
            "minimum": 0.0,
            "title": "Crs Arr Time",
            "description": "Scheduled arrival time, HHMM format. 2400 is accepted as midnight."
          },
          "crs_elapsed_time": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Crs Elapsed Time",
            "description": "Scheduled flight duration in minutes."
          },
          "distance": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Distance",
            "description": "Great-circle distance in miles."
          },
          "flight_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flight Date",
            "description": "Optional ISO scheduled date for exact calendar features."
          }
        },
        "type": "object",
        "required": [
          "airline",
          "origin",
          "destination",
          "month",
          "day_of_week",
          "crs_dep_time",
          "crs_arr_time",
          "crs_elapsed_time",
          "distance"
        ],
        "title": "FlightInput",
        "description": "A single scheduled flight, described using only pre-flight information.",
        "example": {
          "airline": "DL",
          "crs_arr_time": 2145,
          "crs_dep_time": 1830,
          "crs_elapsed_time": 375,
          "day_of_week": 5,
          "destination": "LAX",
          "distance": 2475,
          "flight_date": "2026-07-18",
          "month": 7,
          "origin": "JFK"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PredictionOutput": {
        "properties": {
          "delay_probability": {
            "type": "number",
            "title": "Delay Probability",
            "description": "Post-hoc calibrated estimate of a 15+ minute arrival delay."
          },
          "raw_model_score": {
            "type": "number",
            "title": "Raw Model Score",
            "description": "Uncalibrated score emitted by the selected classifier."
          },
          "calibration_method": {
            "type": "string",
            "title": "Calibration Method",
            "description": "Post-hoc calibration method applied to the raw score."
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level",
            "description": "'low', 'moderate', or 'high'."
          },
          "decision_threshold": {
            "type": "number",
            "title": "Decision Threshold",
            "description": "Probability threshold used for binary risk decisions."
          },
          "top_factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Factors",
            "description": "Non-causal schedule-context signals for this flight."
          },
          "local_contributions": {
            "items": {
              "$ref": "#/components/schemas/LocalContributionOutput"
            },
            "type": "array",
            "title": "Local Contributions",
            "description": "Signed local pre-calibration log-odds contributions from the selected model."
          },
          "explanation_scale": {
            "type": "string",
            "title": "Explanation Scale",
            "description": "Scale used by local model contributions.",
            "default": "log_odds_before_calibration"
          },
          "review_recommended": {
            "type": "boolean",
            "title": "Review Recommended",
            "description": "Whether the configured operational policy sends this flight to priority review.",
            "default": false
          },
          "operational_action": {
            "type": "string",
            "title": "Operational Action",
            "default": "standard_monitoring"
          },
          "policy_name": {
            "type": "string",
            "title": "Policy Name",
            "default": "fixed_threshold"
          },
          "policy_probability_cutoff": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Probability Cutoff"
          },
          "policy_capacity_fraction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Capacity Fraction"
          }
        },
        "type": "object",
        "required": [
          "delay_probability",
          "raw_model_score",
          "calibration_method",
          "risk_level",
          "decision_threshold",
          "top_factors"
        ],
        "title": "PredictionOutput"
      },
...
    }
  }
}